A lot has changed.
First, I added a rough cut at the FSM for the Logic Manager class.
Next, I began to implement a lot of new functionality in the Player and Logic Manager classes that will enable the player to walk around and run into things. I didn't get it completely finished yet, so it definitely won't work or compile, but I had to stop for the night because it's so late.
Also, I changed the sprites.xml and graphics assets to remove some duplicates and to generate the sprites correctly.